requirements.txt
yfinance 
python-telegram-bot 
pandas 
numpy 
ta 
